Resolve circular imports and fix tests with cadence
